The Great Ajax SEO Saga
For the last couple of months I have been working on getting to the
bottom of properly SEO optimising a site that is a fully Javascript
front end.  We thought it would be a simple fun job which in the end
has turned up some very interesting and at the time rather irritating
facts.